refactor: bind 코드 TickerListViewModel로 분리
