+++ CrypMulator/CrypMulator/Present/Ticker/ViewModel/TickerViewModel.swift
@@ -5,4 +5,21 @@
 //  Created by 김태형 on 3/25/25.
 //
 
-import Foundation
+import RxCocoa
+import RxSwift
+
+final class TickerViewModel: ViewModel {
+
+    struct Input {
+
+    }
+
+    struct Output {
+
+    }
+
+
+    func transform(input: Input) -> Output {
+        return Output()
+    }
+}
